Finish project with jwt-spring securoty of GG
You can use and extend by changing the table names or modifying the table values. Instead of many-to-many with Role, you can change it to RolePermission and UserPermission. If you are already using RolePermission and UserPermission, make sure to update the Spring Security UserDetails and the loadByUsername method in UserDetailsService.

In the Spring Security configuration class, there are two extended objects with the same functionality for filtering (AuthTokenFilter - JWTFilter). Remove one and use the other.

Learn about GenericFilterBean and OncePerRequestFilter to understand how to use them effectively. TokenProvider and TokenProviderDuc are essentially the same, so there’s no issue.